B1.3.1
Slight code optimizations in some functions

Added a confirmation warning when trying to close the tool with processes still running

Added "Console" that allows users to execute commands for adb.exe and fastboot.exe manually, printing their outputs in a simple organized interface that also allows copying of the content

Updated the app_manager plugin that corrects some icon extraction errors on devices with customization layers

Fixed a reported bug that caused the tool to terminate execution suddenly after a few uses of App Manager.

App Manager now allows you to display at the bottom of the application list the total number of apps and the number of apps to uninstall or install by selection.