B1.3.3
A "bug" was corrected when loading some decompressed files with 7z (They were not detected as files due to lack of attributes and therefore, the tool ignored them). This new file detection applied almost all functions.

Fixed a bad implementation when updating special file keys manually (Ex: If you changed the destination partition "ZIP FILE" to "UPDATE FILE" the tool only updated the value literally and did not change the file category). Now even if you select a new file from Installation Manager it will try to identify its category automatically or allow it to be changed manually.

Official support was added to the installation of ZIPs for fastboot (UPDATE FILE) and sideload (SIDELOAD FILE). In addition to an automatic generic detection for these.

Added a drop-down list with special file keys (in case automatic file detection is not enough) in the entry of destination partitions

Installation Manager is now resizable.