B1.4.0
Module support! Remote modules can now be downloaded that provide additional logic/functions. All of this is integrated at the interface level so it is quite simple. Additionally, there is support to allow other developers to publish their modules here, but they must consult with me to provide remote linking (with a JSON). Automatic update checking and more.

The File_Manager function now returns an array of the last files/folders selected before closing the window. This can be used by ConfigScripts to ask the user to select a file from the device.

Made some fixes that ensure a clean environment for ConfigScripts